chore: upgrade example to rn-0.74.0-rc.2

+++ android/src/main/old-arch/com/rnmapbox/rnmbx/NativeRNMBXImageModuleSpec.java
@@ -36,5 +36,5 @@ public abstract class NativeRNMBXImageModuleSpec extends ReactContextBaseJavaMod
 
   @ReactMethod
   @DoNotStrip
-  public abstract void refresh(@Nullable Double viewRef, Promise promise);
+  public abstract void refresh(@Nullable Integer viewRef, Promise promise);
 }
